+++ PostHog/PostHogContext.swift
@@ -203,6 +203,7 @@ class PostHogContext {
     /// This centralizes the logic for determining which properties should be used as person properties.
     func personPropertiesContext() -> [String: Any] {
         let staticCtx = staticContext()
+        let sdkInfo = sdkInfo()
         var personProperties: [String: Any] = [:]
 
         // App information
@@ -211,6 +212,10 @@ class PostHogContext {
         }
         if let appBuild = staticCtx["$app_build"] {
             personProperties["$app_build"] = appBuild
+        }
+
+        if let appNamespace = staticCtx["$app_namespace"] {
+            personProperties["$app_namespace"] = appNamespace
         }
 
         // Operating system information
@@ -225,23 +230,25 @@ class PostHogContext {
         if let deviceType = staticCtx["$device_type"] {
             personProperties["$device_type"] = deviceType
         }
-        if let deviceManufacturer = staticCtx["$device_manufacturer"] {
-            personProperties["$device_manufacturer"] = deviceManufacturer
-        }
-        if let deviceModel = staticCtx["$device_model"] {
-            personProperties["$device_model"] = deviceModel
-        }
+//        if let deviceManufacturer = staticCtx["$device_manufacturer"] {
+//            personProperties["$device_manufacturer"] = deviceManufacturer
+//        }
+//        if let deviceModel = staticCtx["$device_model"] {
+//            personProperties["$device_model"] = deviceModel
+//        }
+//
+//        // Localization - read directly to avoid expensive dynamicContext call
+//        if #available(iOS 16.0, macOS 13.0, tvOS 16.0, watchOS 9.0, *) {
+//            if let languageCode = Locale.current.language.languageCode {
+//                personProperties["$locale"] = languageCode.identifier
+//            }
+//        } else {
+//            if let languageCode = Locale.current.languageCode {
+//                personProperties["$locale"] = languageCode
+//            }
+//        }
 
-        // Localization - read directly to avoid expensive dynamicContext call
-        if #available(iOS 16.0, macOS 13.0, tvOS 16.0, watchOS 9.0, *) {
-            if let languageCode = Locale.current.language.languageCode {
-                personProperties["$locale"] = languageCode.identifier
-            }
-        } else {
-            if let languageCode = Locale.current.languageCode {
-                personProperties["$locale"] = languageCode
-            }
-        }
+        personProperties.merge(sdkInfo) { _, new in new }
 
         return personProperties
     }