+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Next
 
+- fix: call the flags api with the correct groups key name (the api has a back compatible fix already) ([#388](https://github.com/PostHog/posthog-ios/pull/388))
+- fix: only set getDefaultPersonProperties with person properties that are evaluated on the server ([#388](https://github.com/PostHog/posthog-ios/pull/388))
+- fix: set and reset PersonPropertiesForFlags and GroupPropertiesForFlags reload flags automatically (or set reloadFeatureFlags: false) ([#388](https://github.com/PostHog/posthog-ios/pull/388))
+
 ## 3.34.0 - 2025-10-15
 
 - feat: add config option to disable swizzling ([#388](https://github.com/PostHog/posthog-ios/pull/388))